requirements.txt
torch
torchvision
pandas
pillow
PyYAML
google-generativeai
pytesseract